Use ``size_t`` for length of a ``std::string``

+++ src/mvcol.cc
@@ -6,7 +6,7 @@
 
 boost::optional< std::shared_ptr<change> > mvcol(contents& contents, boost::optional<int> col) {
     if(col) {
-        unsigned int len = contents.cont[contents.y].length();
+        size_t len = contents.cont[contents.y].length();
         if(len >= static_cast<decltype(len)>(col.get())) {
             contents.x = col.get();
             contents.waiting_for_desired = false;
